refactor: reduce code
