+++ src/app/(normal)/post/[id]/page.tsx
@@ -5,7 +5,7 @@ import { PostDetailPage } from '@/app/(normal)/post/[id]/pages/DetailPage.js';
 import { KeyType, type SocialSourceInURL } from '@/constants/enum.js';
 import { createSiteMetadata } from '@/helpers/createSiteMetadata.js';
 import { isBotRequest } from '@/helpers/isBotRequest.js';
-import { isSocialSourceInURL } from '@/helpers/isSocialSourceInURL.js';
+import { isSocialSourceInURL } from '@/helpers/isSocialSource.js';
 import { memoizeWithRedis } from '@/helpers/memoizeWithRedis.js';
 import { getPostOGById } from '@/services/getPostOGById.js';
 
@@ -21,7 +21,7 @@ interface Props {
 }
 
 export async function generateMetadata({ params, searchParams }: Props): Promise<Metadata> {
-    if (searchParams.source && isSocialSourceInURL(searchParams.source)) {
+    if (isSocialSourceInURL(searchParams.source)) {
         return getPostOGByIdRedis(searchParams.source, params.id);
     }
     return createSiteMetadata();
