+++ src/helpers/isSocialSource.ts
@@ -1,5 +1,14 @@
 import { type SocialSource, Source } from '@/constants/enum.js';
+import { type SocialSourceInURL, SourceInURL } from '@/constants/enum.js';
+import { SORTED_SOCIAL_SOURCES } from '@/constants/index.js';
+import { resolveSourceInURL } from '@/helpers/resolveSourceInURL.js';
 
-export function isSocialSource(source: Source): source is SocialSource {
-    return [Source.Farcaster, Source.Lens, Source.Twitter].includes(source);
+export function isSocialSource(source?: Source): source is SocialSource {
+    if (!source) return false;
+    return SORTED_SOCIAL_SOURCES.includes(source as SocialSource);
+}
+
+export function isSocialSourceInURL(sourceInUrl?: SourceInURL): sourceInUrl is SocialSourceInURL {
+    if (!sourceInUrl) return false;
+    return SORTED_SOCIAL_SOURCES.map(resolveSourceInURL).includes(sourceInUrl);
 }
